compile #5
/home/kde4/svn/_src/koffice/krita/image/tests/kis_painterly_overlay_test.cpp:37: undefined reference 
to `bool QTest::qCompare<unsigned int, unsigned long>(unsigned int const&, unsigned long const&, char 
const*, char const*, char const*, int)'


svn path=/trunk/koffice/; revision=685883

+++ krita/image/tests/kis_painterly_overlay_test.cpp
@@ -34,7 +34,7 @@ void KisPainterlyOverlayTester::testPainterlyOverlayColorSpace()
 {
     KisPainterlyOverlayColorSpace * cs = KisPainterlyOverlayColorSpace::instance();
     Q_ASSERT( cs );
-    QCOMPARE( cs->pixelSize(), 9 * sizeof( float ) );
+    QCOMPARE( cs->pixelSize(), uint(9 * sizeof( float )) );
 }
 
 void KisPainterlyOverlayTester::testPainterlyOverlayColorSpaceCell()
